*: support send restore rbac message
