--- core/restore/secondary/rbac_task.go
@@ -1,4 +1,0 @@
-package secondary
-
-type rbacTask struct {
-}
